# platforms/ycai/studio/docs/openapi-validated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "youchoose ai",
    "description": "youchoose api description",
    "version": "2.3.0"
  },
  "servers": [
    {
      "url": "{protocol}://{host}{port}{basePath}",
      "description": "Node Server",
      "variables": {
        "protocol": {
          "default": "https"
        },
        "host": {
          "default": "api.youchoose.ai"
        },
        "port": {
          "default": "443",
          "enum": ["443", "443"]
        },
        "basePath": {
          "default": ""
        }
      }
    }
  ],
  "security": [
    {
      "ACTToken": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ACTToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Authorization"
      }
    },
...
  },
  "paths": {
...
    "/v3/experiment/{testTime}": {
      "delete": {
        "summary": "ConcludeExperiment",
        "description": "DELETE /v3/experiment/:testTime",
        "tags": ["v3 - Public"],
        "parameters": [
          {
            "name": "testTime",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "string"
            }
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "ConcludeGuardoniExperimentOutput",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "ConcludeGuardoniExperimentOutput",
                  "properties": {
                    "acknowledged": {
                      "type": "boolean",
                      "description": "A valid boolean type"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v3/directives/{experimentId}": {
      "get": {
        "summary": "GetDirective",
        "description": "GET /v3/directives/:experimentId",
        "tags": ["v3 - Public"],
        "parameters": [
          {
            "name": "experimentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "string"
            }
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "GetDirectiveOutput",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": [
                    {
                      "type": "object",
                      "description": "ChiaroScuroDirective",
                      "properties": {
                        "loadFor": {
                          "type": "number",
                          "format": "integer32"
                        },
                        "url": {
                          "type": "string",
                          "description": "string"
                        },
                        "watchFor": {
                          "type": "object",
                          "description": "(number | string | undefined)",
                          "oneOf": [
                            {
                              "type": "number",
                              "format": "integer32"
                            },
                            {
                              "type": "string",
                              "description": "string"
                            }
                          ],
                          "required": false
                        },
                        "name": {
                          "type": "string",
                          "description": "string"
                        },
                        "targetVideoId": {
                          "type": "string",
                          "description": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "description": "ComparisonDirective",
                      "properties": {
                        "title": {
                          "type": "string",
                          "description": "string",
                          "required": false
                        },
                        "url": {
                          "type": "string",
                          "description": "string"
                        },
                        "urltag": {
                          "type": "string",
                          "description": "string",
                          "required": false
                        },
                        "watchFor": {
                          "type": "object",
                          "description": "(number | string | undefined | null)",
                          "oneOf": [
                            {
                              "type": "number",
                              "format": "integer32"
                            },
                            {
                              "type": "string",
                              "description": "string"
                            }
                          ],
                          "required": false
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/v3/handshake": {
      "post": {
        "summary": "Handshake",
        "description": "POST /v3/handshake",
        "tags": ["v3 - Public"],
        "parameters": [],
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "HandshakeBody",
                "properties": {
                  "config": {
                    "type": "object",
                    "description": "{| publicKey: string |}",
                    "properties": {
                      "publicKey": {
                        "type": "string",
                        "description": "string"
                      }
                    }
                  },
                  "href": {
                    "type": "string",
                    "description": "string"
                  },
                  "experimentId": {
                    "type": "string",
                    "description": "string"
                  },
                  "evidencetag": {
                    "type": "string",
                    "description": "string"
                  },
                  "execount": {
                    "type": "string",
                    "description": "string"
                  },
                  "newProfile": {
                    "type": "string",
                    "description": "string"
                  },
                  "testTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "directiveType": {
                    "type": "string",
                    "description": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "any",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "any value"
                }
              }
            }
          }
        }
      }
    },
    "/v3/directives/{directiveType}": {
      "post": {
        "summary": "PostDirective",
        "description": "POST /v3/directives/:directiveType",
        "tags": ["v3 - Public"],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "description": "string"
            }
          },
          {
            "name": "directiveType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "DirectiveType",
              "enum": ["comparison", "chiaroscuro"]
            }
          }
        ],
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "CreateDirectiveBody",
                "properties": {
                  "parsedCSV": {
                    "type": "object",
                    "description": "DirectiveRow",
                    "oneOf": [
                      {
                        "type": "array",
                        "description": "ComparisonDirectiveRow[]",
                        "items": [
                          {
                            "type": "object",
                            "description": "ComparisonDirectiveRow",
                            "properties": {
                              "title": {
                                "type": "string",
                                "description": "string"
                              },
                              "url": {
                                "type": "string",
                                "description": "string"
                              },
                              "urltag": {
                                "type": "string",
                                "description": "string"
                              },
                              "watchFor": {
                                "type": "object",
                                "description": "(number | string | undefined)",
                                "oneOf": [
                                  {
                                    "type": "number",
                                    "format": "integer32"
                                  },
                                  {
                                    "type": "string",
                                    "description": "string"
                                  }
                                ],
                                "required": false
                              }
                            }
                          }
                        ]
                      },
                      {
                        "type": "array",
                        "description": "ChiaroScuroDirectiveRow[]",
                        "items": [
                          {
                            "type": "object",
                            "description": "ChiaroScuroDirectiveRow",
                            "properties": {
                              "videoURL": {
                                "type": "string",
                                "description": "string"
                              },
                              "title": {
                                "type": "string",
                                "description": "string"
                              }
                            }
                          }
                        ]
                      }
                    ],
                    "required": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "PostDirectiveResponse",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "PostDirectiveResponse",
                  "oneOf": [
                    {
                      "type": "object",
                      "description": "{ error: { message: string } }",
                      "properties": {
                        "error": {
                          "type": "object",
                          "description": "{ message: string }",
                          "properties": {
                            "message": {
                              "type": "string",
                              "description": "string"
                            }
                          }
                        }
                      }
                    },
                    {
                      "type": "object",
                      "description": "PostDirectiveSuccessResponse",
                      "properties": {
                        "status": {
                          "type": "string",
                          "description": "(\"exist\" | \"created\")",
                          "enum": ["exist", "created"],
                          "required": true
                        },
                        "experimentId": {
                          "type": "string",
                          "description": "string"
                        },
                        "since": {
                          "type": "object",
                          "description": "(DateFromISOString | undefined)",
                          "oneOf": [
                            {
                              "type": "string",
                              "format": "date-time"
                            }
                          ],
                          "required": false
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}